fix: generateMetaData 라우팅 경로 수정
