+++ app/posts/[postId]/page.tsx
@@ -5,7 +5,6 @@ import styles from "./styles/page.module.scss";
 import { CommentBox } from "./components/clientside";
 import { getPost, mdParser } from "./utils";
 import { CardLayout } from "app/components/card";
-import type { Metadata } from "next";
 
 export default async function Posts({
   params,
@@ -69,49 +68,4 @@ export async function generateStaticParams() {
   }
 
   return staticData;
-}
-
-export async function generateMetadata({
-  params,
-}: {
-  params: { postId: string };
-}): Promise<Metadata> {
-  let URL = process.env.DEV_URL;
-
-  if (typeof URL === undefined) {
-    URL = "https://chocoham.dev";
-  }
-
-  const { postId } = params;
-  const res = await fetch(`${URL}/api/metadata/${postId}`, {
-    method: "GET",
-  });
-  const { data, success } = await res.json();
-
-  if (!success) {
-    throw new Error(data);
-  }
-  const metaData = data[0];
-
-  return {
-    title: metaData.title,
-    description: metaData.description,
-    keywords: metaData.hashtag,
-    bookmarks: [`${URL}/posts/${postId}`],
-    openGraph: {
-      title: metaData.title,
-      description: metaData.description,
-      url: `${URL}/posts/${postId}`,
-      siteName: "ChocoHam 개발 블로그",
-      images: [{ url: metaData.thumbnail, width: 380, height: 250 }],
-      type: "website",
-    },
-    twitter: {
-      card: "summary_large_image",
-      title: metaData.title,
-      description: metaData.description,
-      creator: "초코햄",
-      images: [metaData.thumbnail],
-    },
-  };
 }
